LPS-74877 Move to content processor package
