+++ adaptive-media-image-service/src/main/java/com/liferay/adaptive/media/image/internal/exportimport/content/processor/AMEmbeddedReferenceSetFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.image.internal.exportimport;
+package com.liferay.adaptive.media.image.internal.exportimport.content.processor;
 
 import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.exportimport.kernel.lar.ExportImportPathUtil;
